Talking Medbots!
Changes some desctexts and makes Medbots talk to ya! They'll announce
aloud when they're approaching someone to give 'em the good stuff, and
will comment on the nature and severity of their injuries, where voice
line availability allows.

I have bigger plans than just this, including enabling them to comment
on major/minor fracture status and internal bleeding, as well as
advising their patient when they're gonna need a real doctor for THAT
boo-boo, and many of these may be tied in to a general overhaul of
medbots in general. But, each big adventure has a first step, and I'm
happy with this in its own right.