Emagging Fixed
This should be the final tweak for this package. The Medbots Ai now
works correctly once they've been emagged, and I've tweaked what they
inject, as well as how much, to make them a credible threat but not an
overwhelming force.

They still have the issue of their weird behaviour giving them away
almost instantly, but unless they suddenly develop the power to emit
invisible toxic gas or something, that's always going to be an issue.